server/requirements.txt
huggingface-hub>=0.19.0
python-dotenv>=1.0.0
gtts>=2.4.0
requests>=2.31.0